Altercao upload arquivos

+++ src/middlewares/resizeFile.js
@@ -4,17 +4,21 @@ const fs = require('fs')
 
 
 const resizeFile = (req, res, next) => {
-    const { filename: image } = req.file
-    sharp(req.file.path)
-    .resize(89,93)
-    .jpeg({quality: 50})
-    .toFile(
-            path.resolve(`${req.file.destination}`,'resized',image)
-    )
-    .then(data => {
-        fs.unlinkSync(req.file.path)
+    if(req.file){
+        const { filename: image } = req.file
+        sharp(req.file.path)
+        .resize(89,93)
+        .jpeg({quality: 50})
+        .toFile(
+                path.resolve(`${req.file.destination}`,'resized',image)
+        )
+        .then(data => {
+            fs.unlinkSync(req.file.path)
+            next()
+        })
+    }else{
         next()
-    })
+    }
     
 }
